fix bug where EntityGetEndpoint would hang if items/orderedItems was not an array
as2 allows items/orderedItems to be a Link object, not just an array, so this adds a check on the value before mapping over it. Without this, a non-array value would lead to an uncaught exception.

related, activitypub-core-server-express should probably catch unexpected exceptions in the endpoint `respond` methods and pass the error to express's `next(err)` callback. without that, express will hang on uncaught errors because `res.end()` is never called